+++ frontend/package.json
@@ -11,6 +11,8 @@
     "@react-native-async-storage/async-storage": "^1.13.2",
     "@react-native-community/datetimepicker": "3.0.0",
     "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/bottom-tabs": "^5.11.1",
+    "@react-navigation/material-bottom-tabs": "^5.3.9",
     "@react-navigation/native": "^5.7.6",
     "@react-navigation/stack": "^5.9.3",
     "axios": "^0.21.0",
@@ -24,14 +26,15 @@
     "react-native-gesture-handler": "~1.7.0",
     "react-native-maps": "^0.27.1",
     "react-native-modal-datetime-picker": "^9.0.0",
-    "react-native-paper": "^4.2.0",
+    "react-native-paper": "^4.4.0",
     "react-native-paper-datepicker": "^1.0.0",
     "react-native-qrcode-svg": "^6.0.6",
     "react-native-reanimated": "~1.13.0",
     "react-native-safe-area-context": "3.1.4",
     "react-native-screens": "~2.10.1",
     "react-native-svg": "12.1.0",
-    "react-native-web": "~0.13.12"
+    "react-native-web": "~0.13.12",
+    "react-native-webview": "10.7.0"
   },
   "devDependencies": {
     "@babel/core": "~7.9.0"
